+++ src/main/kotlin/com/github/jozott00/wokwiintellij/actions/WokwiStopAction.kt
@@ -1,7 +1,7 @@
 package com.github.jozott00.wokwiintellij.actions
 
 
-import com.github.jozott00.wokwiintellij.services.WokwiProjectService
+import com.github.jozott00.wokwiintellij.services.WokwiSimulatorService
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -14,15 +14,14 @@ class WokwiStopAction : AnAction() {
     }
 
     override fun update(e: AnActionEvent) {
-        val s = e.project?.service<WokwiProjectService>() ?: return
+        val s = e.project?.service<WokwiSimulatorService>() ?: return
         val p = e.presentation
 
         p.isEnabled = s.isSimulatorRunning()
-
     }
 
     override fun actionPerformed(event: AnActionEvent) {
-        val s = event.project?.service<WokwiProjectService>()
+        val s = event.project?.service<WokwiSimulatorService>()
         s?.stopSimulator()
 
     }