+++ src/main/kotlin/com/github/jozott00/wokwiintellij/simulator/WokwiSimulatorListener.kt
@@ -5,7 +5,12 @@ import com.intellij.openapi.util.Key
 
 interface WokwiSimulatorListener {
     fun onStarted(runArgs: WokwiArgs) {}
-    fun onShutdown() {}
+    fun onShutdown(exitCode: SimExitCode) {}
     fun onTextAvailable(text: String, outputType: Key<*>) {}
     fun onRunning() {}
+}
+
+enum class SimExitCode(val int: Int) {
+    OK(0),
+    CONFIG_ERROR(1),
 }