+++ CHANGELOG.md
@@ -3,8 +3,27 @@
 # wokwi-intellij Changelog
 
 ## [Unreleased]
+
+This update switches from the custom run toolwindow
+to a headless run configuration, which is more familiar for
+most Intellij users.
+
 ### Added
+- Run configuration for simulator instance
+- Automatic GDB port selection
+
+### Removed
+- WokwiDebugStart configuration
+- Wokwi Run toolwindow
+
+## [0.9.0] - 2024-02-23
+
+### Added
+
 - Embedded Wokwi simulator
 - Wokwi console to control simulator
 - wokwi.toml analysis support
 - Wokwi simulation debugging support
+
+[Unreleased]: https://github.com/Jozott00/wokwi-intellij/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/Jozott00/wokwi-intellij/commits/v0.9.0